@@ -1,6 +1,6 @@
 {
   "name": "game-roles-v2",
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "test": "src/tests",
   "license": "AGPL-3.0",
   "author": "tippfehlr",
@@ -9,24 +9,22 @@
     "url": "https://github.com/tippf3hlr/game-roles-v2.git"
   },
   "dependencies": {
-    "discord.js": "^12.5.1",
+    "bufferutil": "^4.0.3",
+    "discord.js": "^13.1.0",
     "dotenv": "^10.0.0",
-    "mongoose": "^5.12.13",
-    "node": "^16.6.1",
     "nodemon": "^2.0.12",
-    "wokcommands": "^1.3.6"
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.2",
+    "utf-8-validate": "^5.0.5",
+    "wokcommands": "^1.3.9303"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
-    "chai": "^4.3.4",
-    "eslint": "^7.29.0",
-    "mocha": "^9.0.0"
+    "eslint": "^7.32.0"
   },
   "scripts": {
-    "pm2": "pm2 start ./src/index.js -a --restart-delay 3000 --name game-roles-v2",
-    "run": "node ./src/index.js",
-    "nodemon": "nodemon",
+    "pm2": "pm2 start ./src/index.ts -a --restart-delay 3000 --name game-roles-v2",
+    "ts-node": "node ./src/index.ts",
+    "dev": "nodemon",
     "test": "mocha"
   }
 }
